requirements.txt
pandas 
google-cloud-pubsub 
pypi-json
uuid